requirements.txt
Django==4.2.15
celery==5.4.0
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
djangorestframework==3.13.1
django-environ==0.8.1
pillow==10.4.0
gunicorn==20.1.0
psycopg2==2.9.2
psycopg2-binary==2.9.2